+++ python/ray/llm/_internal/serve/deployments/routers/middleware.py
@@ -8,7 +8,7 @@ from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoin
 from starlette.responses import JSONResponse, Response
 
 from ray.llm._internal.serve.observability.logging import get_logger
-from ray.llm._internal.serve.deployments.server_utils import (
+from ray.llm._internal.serve.deployments.utils.server_utils import (
     get_response_for_error,
 )
