+++ python/ray/llm/_internal/serve/deployments/utils/error_handling_utils.py
@@ -8,9 +8,9 @@ from ray.llm._internal.serve.observability.logging import get_logger
 from ray.llm._internal.serve.observability.metrics.utils import (
     InstrumentTokenAsyncGenerator,
 )
-from ray.llm._internal.serve.deployments.metrics import Metrics
+from ray.llm._internal.serve.deployments.utils.metrics_utils import Metrics
 from ray.llm._internal.serve.configs.server_models import LLMRawResponse
-from ray.llm._internal.serve.deployments.server_utils import (
+from ray.llm._internal.serve.deployments.utils.server_utils import (
     get_response_for_error,
     get_serve_request_id,
 )
