+++ python/ray/llm/tests/serve/deployments/mock_vllm_engine.py
@@ -25,7 +25,7 @@ from ray.llm._internal.serve.deployments.llm.vllm.vllm_models import (
     VLLMGenerationRequest,
     VLLMSamplingParams,
 )
-from ray.llm._internal.serve.deployments.llm_node_initializer import (
+from ray.llm._internal.serve.deployments.utils.node_initialization_utils import (
     InitializeNodeOutput,
 )
